Order by relational aggregates
